requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
anthropic==0.52.0
openai==1.82.0
python-dotenv==1.1.0
pdfplumber==0.11.6
python-docx==1.1.2
